+++ .config/nvim/.luarc.json
@@ -1,0 +1,5 @@
+{
+    "diagnostics.disable": [
+        "undefined-field"
+    ]
+}
