+++ .config/nvim/.neoconf.json
@@ -13,4 +13,3 @@
     }
   }
 }
-
