--- .config/nvim/lua/plugins/configs/nvim-lsp-installer.lua
@@ -1,10 +1,0 @@
-local status_ok, package = pcall(require, "nvim-lsp-installer")
-if not status_ok then
-	return
-end
-
-package.setup({
-	ui = {
-		border = "double",
-	},
-})
