--- .config/nvim/lua/plugins/configs/nvim-scrollbar.lua
@@ -1,92 +1,0 @@
-local status_ok, scrollbar = pcall(require, "scrollbar")
-if not status_ok then
-	return
-end
-
-scrollbar.setup({
-	show = true,
-	show_in_active_only = false,
-	set_highlights = true,
-	folds = 1000, -- handle folds, set to number to disable folds if no. of lines in buffer exceeds this
-	max_lines = false, -- disables if no. of lines in buffer exceeds this
-	handle = {
-		text = " ",
-		color = nil,
-		cterm = nil,
-		highlight = "CursorColumn",
-		hide_if_all_visible = true, -- Hides handle if all lines are visible
-	},
-	marks = {
-		Search = {
-			text = { "-", "=" },
-			priority = 0,
-			color = nil,
-			cterm = nil,
-			highlight = "Search",
-		},
-		Error = {
-			text = { "-", "=" },
-			priority = 1,
-			color = "#d18686", -- red
-			cterm = nil,
-			highlight = "DiagnosticVirtualTextError",
-		},
-		Warn = {
-			text = { "-", "=" },
-			priority = 2,
-			color = "#f9ab00", -- yellow
-			cterm = nil,
-			highlight = "DiagnosticVirtualTextWarn",
-		},
-		Info = {
-			text = { "-", "=" },
-			priority = 3,
-			color = "#bb86d1", -- purple
-			cterm = nil,
-			highlight = "DiagnosticVirtualTextInfo",
-		},
-		Hint = {
-			text = { "-", "=" },
-			priority = 4,
-			color = "#86d195", -- green
-			cterm = nil,
-			highlight = "DiagnosticVirtualTextHint",
-		},
-		Misc = {
-			text = { "-", "=" },
-			priority = 5,
-			color = nil,
-			cterm = nil,
-			highlight = "Normal",
-		},
-	},
-	excluded_buftypes = {
-		"terminal",
-	},
-	excluded_filetypes = {
-		"prompt",
-		"TelescopePrompt",
-	},
-	autocmd = {
-		render = {
-			"BufWinEnter",
-			"TabEnter",
-			"TermEnter",
-			"WinEnter",
-			"CmdwinLeave",
-			"TextChanged",
-			"VimResized",
-			"WinScrolled",
-		},
-		clear = {
-			"BufWinLeave",
-			"TabLeave",
-			"TermLeave",
-			"WinLeave",
-		},
-	},
-	handlers = {
-		diagnostic = true,
-		search = false, -- Requires hlslens to be loaded, will run require("scrollbar.handlers.search").setup() for you
-	},
-})
