+++ .config/nvim/lua/plugins/lsp.lua
@@ -39,4 +39,10 @@ return {
       require("plugins.lsp.handlers").setup()
     end,
   },
+  {
+    "ray-x/lsp_signature.nvim",
+    event = "VeryLazy",
+    opts = {},
+    config = function(_, opts) require 'lsp_signature'.setup(opts) end
+  }
 }
