+++ .config/nvim/lua/plugins/mason.lua
@@ -2,6 +2,10 @@ return {
   "williamboman/mason.nvim",
   lazy = false,
   config = function()
-    require('mason').setup()
+    require('mason').setup({
+      ui = {
+        border = "double"
+      }
+    })
   end
 }
